Entered Main and made ancillary change
Adventure main is playtest complete and bug free. But I'm open to making
extensions and will be cleaning it up once matt sends me the
descriptions. Consider the text placeholders.

Also, added "from media import *" to player, inputParser and room as it
was required to run. Lastly, changed the true on line 20 of inputParser
to True due to a strange bug.